Refactoring management of approxData instances and model keys: > Approximation::approxData now consists only of shallow copies of   Pecos::PolynomialApproximation::{surrData,modSurrData}.  These two approx   data sets now utilize more careful management of keys, allowing multiple   data sets that comprise discrepancy pairings to co-exist and be properly   distinguished. > Consolidation of SizetSizetPair model indices with UShortArray model keys   throughout Model, Interface, Approximation. > Extension of virtual key assignments for truth and surrogate models to allow   more unified handling of aggregated response data within Model recursions. > Flip response aggregation ordering from {LF,HF} to {HF,LF} in order to better   synchronize with key-based management of approx data sets.

+++ src/SimulationModel.hpp
@@ -56,9 +56,9 @@ protected:
   /// of 1 solution level
   size_t solution_levels(bool lwr_bnd = true) const;
   /// activate entry in solnControlCostMap
-  void solution_level_index(size_t lev_index);
+  void solution_level_index(unsigned short lev_index);
   /// return active entry in solnControlCostMap
-  size_t solution_level_index() const;
+  unsigned short solution_level_index() const;
   /// return all cost estimates from solnControlCostMap
   RealVector solution_level_costs() const;
   /// return active cost estimate from solnControlCostMap
